requirements.txt
ollama>=0.2.0
anthropic>=0.25.0
openai>=1.30.0
PyGithub>=2.3.0
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.7
requests>=2.31.0